Changed double to BigDecimal

+++ src/main/kotlin/com/euvsvirus/euvsvirus/workout/api/CreateWorkoutRequest.kt
@@ -1,10 +1,11 @@
 package com.euvsvirus.euvsvirus.workout.api
 
+import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 data class CreateWorkoutRequest(
         val datetimeStart: ZonedDateTime,
         val datetimeEnd: ZonedDateTime,
         val sport: String,
-        val points: List<List<Double>>
+        val points: List<List<BigDecimal>>
 )
